pyproject.toml
[tool.poetry]
name = "wireless-security"
version = "0.1.0"
description = ""
authors = ["MaxMohammadi <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
pycrypto = "^2.6.1"
pycryptodome = "^3.12.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
